// src/__generated__/graphql.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: { input: string; output: string };
  String: { input: string; output: string };
  Boolean: { input: boolean; output: boolean };
  Int: { input: number; output: number };
  Float: { input: number; output: number };
  /** A date-time string at UTC, such as 2007-12-03T10:15:30Z, compliant with the `date-time` format outlined in section 5.6 of the RFC 3339 profile of the ISO 8601 standard for representation of dates and times using the Gregorian calendar. */
  DateTime: { input: any; output: any };
  /** A field whose value conforms with the standard mongodb object Id as described here: https://docs.mongodb.com/manual/reference/method/ObjectId/#ObjectId. Example: 5e5677d71bdc2ae76344968c */
  ObjectId: { input: any; output: any };
};
// ...
export type AudioUrls = {
  __typename?: 'AudioUrls';
  /** Female version of audio url of the post. */
  female: Maybe<Scalars['String']['output']>;
  /** Male version of audio url of the post. */
  male: Maybe<Scalars['String']['output']>;
};
// ...
export type Content = {
  __typename?: 'Content';
  /** The HTML version of the content. */
  html: Scalars['String']['output'];
  /** The Markdown version of the content. */
  markdown: Scalars['String']['output'];
  /** The text version from sanitized html content. HTML tags are stripped and only text is returned. */
  text: Scalars['String']['output'];
};
// ...
export type Node = {
  /** The ID of the node. */
  id: Scalars['ID']['output'];
};
// ...
export type OpenGraphMetaData = {
  __typename?: 'OpenGraphMetaData';
  /** The image used in og:image tag for SEO purposes. */
  image: Maybe<Scalars['String']['output']>;
};
// ...
export type Post = Node & {
  __typename?: 'Post';
  /** Returns male and female audio url of the post. Available in case the Audioblog is enabled. */
  audioUrls: Maybe<AudioUrls>;
  /** Returns the user details of the author of the post. */
  author: User;
  /**
   * Flag to indicate if the post is bookmarked by the requesting user.
   *
   * Returns `false` if the user is not authenticated.
   */
  bookmarked: Scalars['Boolean']['output'];
  /** Brief is a short description of the post extracted from the content of the post. It's 250 characters long sanitized string. */
  brief: Scalars['String']['output'];
  /** Canonical URL set by author in case of republished posts. */
  canonicalUrl: Maybe<Scalars['String']['output']>;
  /**
   * Returns the user details of the co-authors of the post.
   * Hashnode users can add up to 4 co-authors as collaborators to their posts.
   * This functionality is limited to teams publication.
   */
  coAuthors: Maybe<Array<User>>;
  /** List of users who have commented on the post. */
  commenters: PostCommenterConnection;
  /** A list of comments on the post. */
  comments: PostCommentConnection;
  /** Content of the post. Contains HTML and Markdown version of the post content. */
  content: Content;
  /**
   * A list of contributors of the post. Contributors are users who have commented or replied to the post.
   * @deprecated Will be removed on 10th Oct 2023. Use `commenters` instead.
   */
  contributors: Array<User>;
  /** The cover image preference of the post. Contains cover image URL and other details. */
  coverImage: Maybe<PostCoverImage>;
  /** Unique ID to identify post, used internally by hashnode. */
  cuid: Maybe<Scalars['String']['output']>;
  /** Flag to indicate if the post is featured on Hashnode feed. */
  featured: Scalars['Boolean']['output'];
  /** The date and time the post was featured. Used along with featured flag to determine if the post is featured. */
  featuredAt: Maybe<Scalars['DateTime']['output']>;
  /** Post feature-related fields. */
  features: PostFeatures;
  /** A flag to indicate if the post contains LaTeX. Latex is used to write mathematical equations. */
  hasLatexInPost: Scalars['Boolean']['output'];
  /** The ID of the post. Used to uniquely identify the post. */
  id: Scalars['ID']['output'];
  /** Wether or not the post has automatically been published via RSS feed. */
  isAutoPublishedFromRSS: Scalars['Boolean']['output'];
  /**
   * Wether or not the authenticated user is following this post.
   *
   * Returns `null` if the user is not authenticated.
   */
  isFollowed: Maybe<Scalars['Boolean']['output']>;
  /** A list of users who liked the post. */
  likedBy: PostLikerConnection;
  /** OG meta-data of the post. Contains image url used in open graph meta tags. */
  ogMetaData: Maybe<OpenGraphMetaData>;
  /** Preference settings for the post. Contains information about if the post is pinned to blog, comments are disabled, etc. */
  preferences: PostPreferences;
  /** The publication the post belongs to. */
  publication: Maybe<Publication>;
  /** The date and time the post was published. */
  publishedAt: Scalars['DateTime']['output'];
  /** The number of hearts on the post. Shows how many users liked the post. */
  reactionCount: Scalars['Int']['output'];
  /** The estimated time to read the post in minutes. */
  readTimeInMinutes: Scalars['Int']['output'];
  /** The number of replies on the post. */
  replyCount: Scalars['Int']['output'];
  /** The number of comments on the post. */
  responseCount: Scalars['Int']['output'];
  /** SEO information of the post. Contains title and description used in meta tags. */
  seo: Maybe<Seo>;
  /** Information of the series the post belongs to. */
  series: Maybe<Series>;
  /** The slug of the post. Used as address of the post on blog. Example - https://johndoe.com/my-post-slug */
  slug: Scalars['String']['output'];
  /** The subtitle of the post. Subtitle is a short description of the post which is also used in SEO if meta tags are not provided. */
  subtitle: Maybe<Scalars['String']['output']>;
  /** Returns list of tags added to the post. Contains tag id, name, slug, etc. */
  tags: Maybe<Array<Tag>>;
  /** The title of the post. */
  title: Scalars['String']['output'];
  /** The date and time the post was last updated. */
  updatedAt: Maybe<Scalars['DateTime']['output']>;
  /** Complete URL of the post including the domain name. Example - https://johndoe.com/my-post-slug */
  url: Scalars['String']['output'];
  /** The number of views on the post. Can be used to show the popularity of the post. */
  views: Scalars['Int']['output'];
};
// ...
export type PostCommentConnection = Connection & {
  __typename?: 'PostCommentConnection';
  /** A list of edges containing comments as nodes. */
  edges: Array<PostCommentEdge>;
  /** Information about pagination in a connection. */
  pageInfo: PageInfo;
  /** Total number of nodes available i.e. number of comments. */
  totalDocuments: Scalars['Int']['output'];
};
// ...
export type PostCommenterConnection = Connection & {
  __typename?: 'PostCommenterConnection';
  /** A list of edges containing commenters as nodes. */
  edges: Array<PostCommenterEdge>;
  /** Information about pagination in a connection. */
  pageInfo: PageInfo;
  /** Total number of nodes available i.e. number of commenters. */
  totalDocuments: Scalars['Int']['output'];
};
// ...
export type PostCoverImage = {
  __typename?: 'PostCoverImage';
  /** Provides attribution information for the cover image, if available. */
  attribution: Maybe<Scalars['String']['output']>;
  /** True if the image attribution should be hidden. */
  isAttributionHidden: Scalars['Boolean']['output'];
  /** Indicates whether the cover image is in portrait orientation. */
  isPortrait: Scalars['Boolean']['output'];
  /** The name of the photographer who captured the cover image. */
  photographer: Maybe<Scalars['String']['output']>;
  /** The URL of the cover image. */
  url: Scalars['String']['output'];
};

/** Contains a post and a cursor for pagination. */
export type PostEdge = Edge & {
  __typename?: 'PostEdge';
  /** A cursor for use in pagination. */
  cursor: Scalars['String']['output'];
  /** The node holding the Post information */
  node: Post;
};

export type PostFeatures = {
  __typename?: 'PostFeatures';
  badges: PostBadgesFeature;
  tableOfContents: TableOfContentsFeature;
};

/**
 * Connection for users who liked the post. Contains a list of edges containing nodes.
 * Each node is a user who liked the post.
 * Page info contains information about pagination like hasNextPage and endCursor.
 * Total documents contains the total number of users who liked the post.
 */
export type PostLikerConnection = Connection & {
  __typename?: 'PostLikerConnection';
  /** A list of edges containing users as nodes */
  edges: Array<PostLikerEdge>;
  /** Information about pagination in a connection. */
  pageInfo: PageInfo;
  /** Total number of nodes available i.e. number of users who liked the post. */
  totalDocuments: Scalars['Int']['output'];
};
// ...
export type PostPreferences = {
  __typename?: 'PostPreferences';
  /** A flag to indicate if the comments are disabled for the post. */
  disableComments: Scalars['Boolean']['output'];
  /** Wether or not the post is hidden from the Hashnode community. */
  isDelisted: Scalars['Boolean']['output'];
  /** A flag to indicate if the post is pinned to blog. Pinned post is shown on top of the blog. */
  pinnedToBlog: Scalars['Boolean']['output'];
  /** A flag to indicate if the cover image is shown below title of the post. Default position of cover is top of title. */
  stickCoverToBottom: Scalars['Boolean']['output'];
};
// ...
export type User = IUser &
  Node & {
    __typename?: 'User';
    /**
     * Whether or not the user is an ambassador.
     * @deprecated Ambassadors program no longer active. Will be removed after 02/01/2024
     */
    ambassador: Scalars['Boolean']['output'];
    /** The availability of the user based on tech stack and interests. Shown on the "I am available for" section in user's profile. */
    availableFor: Maybe<Scalars['String']['output']>;
    /** Returns a list of badges that the user has earned. Shown on blogs /badges page. Example - https://iamshadmirza.com/badges */
    badges: Array<Badge>;
    /** The bio of the user. Visible in about me section of the user's profile. */
    bio: Maybe<Content>;
    /**
     * The bio of the user. Visible in about me section of the user's profile.
     * @deprecated Will be removed on 26/10/2023. Use bio instead of bioV2
     */
    bioV2: Maybe<Content>;
    /** The date the user joined Hashnode. */
    dateJoined: Maybe<Scalars['DateTime']['output']>;
    /** Whether or not the user is deactivated. */
    deactivated: Scalars['Boolean']['output'];
    /** The users who are following this user */
    followers: UserConnection;
    /** The number of users that follow the requested user. Visible in the user's profile. */
    followersCount: Scalars['Int']['output'];
    /**
     * Wether or not the authenticated user follows this user.
     * Returns false if the authenticated user this user.
     */
    following: Scalars['Boolean']['output'];
    /** The number of users that this user is following. Visible in the user's profile. */
    followingsCount: Scalars['Int']['output'];
    /** The users which this user is following */
    follows: UserConnection;
    /**
     * Wether or not this user follows the authenticated user.
     * Returns false if the authenticated user this user.
     */
    followsBack: Scalars['Boolean']['output'];
    /** The ID of the user. It can be used to identify the user. */
    id: Scalars['ID']['output'];
    /** Wether or not this is a pro user. */
    isPro: Scalars['Boolean']['output'];
    /** The location of the user. */
    location: Maybe<Scalars['String']['output']>;
    /** The name of the user. */
    name: Scalars['String']['output'];
    /** Returns the list of posts the user has published. */
    posts: UserPostConnection;
    /** The URL to the profile picture of the user. */
    profilePicture: Maybe<Scalars['String']['output']>;
    /** Publications associated with the user. Includes personal and team publications. */
    publications: UserPublicationsConnection;
    /** The social media links of the user. Shown on the user's profile. */
    socialMediaLinks: Maybe<SocialMediaLinks>;
    /** The tagline of the user. Shown on the user's profile below the name. */
    tagline: Maybe<Scalars['String']['output']>;
    /** Returns a list of tags that the user follows. */
    tagsFollowing: Array<Tag>;
    /** The username of the user. It is unique and tied with user's profile URL. Example - https://hashnode.com/@username */
    username: Scalars['String']['output'];
  };